+++ sample-dbschema/sample_schema-less-locking.pgsql.sql
@@ -1,0 +1,14 @@
+-- Sample table schema for less-locking PgSQL queue
+
+DROP TABLE IF EXISTS queuell;
+CREATE TABLE queuell (
+    queue_id                    BIGSERIAL,
+        PRIMARY KEY (queue_id),
+    ephemeral_id                BIGINT              NOT NULL DEFAULT 0,
+    msg_org_timestamp           TIMESTAMP           NOT NULL,
+    msg_timestamp               TIMESTAMP           NOT NULL,
+    msg_num_requeues            INT                 NOT NULL DEFAULT 0,
+    msg_content                 BYTEA
+);
+CREATE INDEX queuell_ephemeral_id ON queuell(ephemeral_id);
+CREATE INDEX queuell_msg_timestamp ON queuell(msg_timestamp);